requirements.txt
aiogram
python-dotenv
transformers
datasets
torch